requirements.txt
requests
subprocess32
python-dotenv
python-gitlab